+++ feature/activity/src/main/java/akio/apps/myrun/feature/activitydetail/ActivityExportService.kt
@@ -2,7 +2,7 @@ package akio.apps.myrun.feature.activitydetail
 
 import akio.apps.myrun.domain.activity.ExportTempTcxFileUsecase
 import akio.apps.myrun.feature.activity.R
-import akio.apps.myrun.feature.activitydetail.wiring.DaggerActivityExportFeatureComponent
+import akio.apps.myrun.feature.activitydetail.di.DaggerActivityExportFeatureComponent
 import akio.apps.myrun.feature.base.AppNotificationChannel
 import android.app.Notification
 import android.app.PendingIntent
