+++ feature/profile/src/main/java/akio/apps/myrun/feature/profile/LinkStravaActivity.kt
@@ -5,7 +5,7 @@ import akio.apps.myrun.feature.base.DialogDelegate
 import akio.apps.myrun.feature.base.ktx.collectEventRepeatOnStarted
 import akio.apps.myrun.feature.base.ktx.collectRepeatOnStarted
 import akio.apps.myrun.feature.base.viewmodel.lazyViewModelProvider
-import akio.apps.myrun.feature.profile.wiring.DaggerLinkStravaComponent
+import akio.apps.myrun.feature.profile.di.DaggerLinkStravaComponent
 import android.content.Intent
 import android.os.Bundle
 import android.widget.Toast
