+++ feature/tracking/src/main/java/akio/apps/myrun/feature/tracking/RouteTrackingActivity.kt
@@ -15,6 +15,7 @@ import akio.apps.myrun.feature.base.ktx.collectRepeatOnStarted
 import akio.apps.myrun.feature.base.ktx.dp2px
 import akio.apps.myrun.feature.base.ktx.observe
 import akio.apps.myrun.feature.base.viewmodel.lazyViewModelProvider
+import akio.apps.myrun.feature.tracking.di.DaggerRouteTrackingFeatureComponent
 import akio.apps.myrun.feature.tracking.ui.ActivitySettingsView
 import akio.apps.myrun.feature.tracking.ui.RouteTrackingStatsView
 import akio.apps.myrun.feature.tracking.ui.StopDialogOptionId
@@ -22,7 +23,6 @@ import akio.apps.myrun.feature.tracking.ui.StopOptionsDialog
 import akio.apps.myrun.feature.tracking.ui.TrackingControlButtonPanel
 import akio.apps.myrun.feature.tracking.ui.TrackingControlButtonType
 import akio.apps.myrun.feature.tracking.utils.LatLngBoundsBuilder
-import akio.apps.myrun.feature.tracking.wiring.DaggerRouteTrackingFeatureComponent
 import akio.apps.myrun.worker.ActivityUploadWorker
 import android.annotation.SuppressLint
 import android.content.Context
